+++ FBSDKShareKit/FBSDKShareKit/FBSDKShareKit.h
@@ -23,7 +23,6 @@
 
 #if !TARGET_OS_TV
  #import <FBSDKShareKit/FBSDKAppInviteContent.h>
- #import <FBSDKShareKit/FBSDKMessageDialog.h>
  #import <FBSDKShareKit/FBSDKShareCameraEffectContent.h>
  #import <FBSDKShareKit/FBSDKShareDialog.h>
  #import <FBSDKShareKit/_FBSDKMessengerIcon.h>